Work around bmread use of uninitialized memory
bmread copies from a stale pointer into a local buffer.  If the stale
pointer points to null bytes, this is stupid but harmless.  If the stale
pointer points to a large run of non-null bytes, this crashes the game.
Modify PHYSFSX_fgets to clear the unused portion of the return buffer to
work around this bmread bug.  Otherwise using shareware data crashes the
game at startup.