Fix __builtin_constant_p test at -Og
Remove explicit 'inline' on __builtin_constant_p test, since -Og will
inline simple functions on request but not inline speculatively or
inline complex templates.  When testing at -Og with 'static inline',
__builtin_constant_p is marked as functional, but physfsx checks are not
inlined and result in spurious failures.  When testing at -Og with
'static', __builtin_constant_p is marked as broken and the physfsx
checks are skipped.  When testing at -O2, __builtin_constant_p is marked
as working and the physfsx checks are included.