Respect autoselect order for granted items
Additionally, improve handling of Primary_last_was_super.  Previously,
it was cleared to 0, which made sense if the ship starts with only
lasers.  Now that the ship can start with other weapons, clearing
Primary_last_was_super to 0 may override a player's choice.  Preserve
player's last-was-super choice when the grant gives both the base and
super forms of a weapon.

Requested-by: Mako88 <https://github.com/dxx-rebirth/dxx-rebirth/issues/140> [for the Primary_weapon part]