Fix lifetime kill accounting in team hoard
`adjust` must be set to -1 for a team kill, but this was only done as a
side effect of updating game-local kill counts.  After 5f2ec67,
game-local kill counts are not updated in team hoard, but adjust is
still used.  Move the handling of adjust up to occur in all team games.

Fixes: 5f2ec6789fbcb5dfcab2fc81c521380d94ac3294 ("Restore !hoard guard for game-local kill accounting")