Hide bogus demo cloak/invul timers
Demo files do not record cloak/invulnerability times, so the playback
code sets a fake time of "50% elapsed" for the entire time that the
effect is active.  This causes two problems in playback.  The first,
which always applied, but was mostly ignored, is that the viewer does
not get the blinking warning that the effect will expire soon.  The
second, which became visible with the addition of cloak/invulnerability
digital timers, is that demos always report "15" as the time remaining,
for the entire time that the effect is active.  Old demos cannot be
fixed, so hide the problem by forcing the timers to be hidden when
viewing a demo.