Add special case to make past releases not die on MarkerObject
All releases to date have a bug where they treat MarkerObject as an int,
not an object number.  Storing object_none (0xffff) into the save file
causes affected releases to exhibit several problems.  Most obviously,
they crash if the automap is opened, because they try to show
Objects[0xffff] as a marker, but no such object exists.  Additionally,
they refuse to let the player drop a marker because none of the elements
of MarkerObject are -1, so they incorrectly tell the player that all
marker slots are busy.  Finally, they will crash if the guidebot is
killed, because the guidebot death code deletes marker 0xffff because it
fails to recognize that this is object_none.

Current code correctly treats object_none as a non-object and works
correctly without this hack.  The hack is only required to get past
releases to work correctly after loading a saved game written by current
code.

Fixes: 9125ae32cdd561dec0983561f7bfaee72c07d788 ("Make objnum unsigned")